+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/map/QuestionMapService.java
@@ -5,7 +5,16 @@ import org.springframework.stereotype.Service;
 import ru.dpopkov.knowthenics.model.Question;
 import ru.dpopkov.knowthenics.services.QuestionService;
 
+import java.util.Set;
+import java.util.stream.Collectors;
+
 @Service
 @Profile({"default", "map-service"})
 public class QuestionMapService extends AbstractMapService<Question> implements QuestionService {
+    @Override
+    public Set<Question> findAllByWordingEnLike(String searchPattern) {
+        return super.findAll().stream()
+                .filter(question -> question.getWordingEn() != null && question.getWordingEn().contains(searchPattern))
+                .collect(Collectors.toSet());
+    }
 }