+++ knowthenics-web/src/main/java/ru/dpopkov/knowthenics/controllers/QuestionController.java
@@ -2,17 +2,25 @@ package ru.dpopkov.knowthenics.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 import ru.dpopkov.knowthenics.model.Question;
 import ru.dpopkov.knowthenics.services.QuestionService;
 
+import java.util.ArrayList;
+import java.util.Set;
+
 @SuppressWarnings("SameReturnValue")
 @Controller
 @RequestMapping("/questions")
 public class QuestionController {
+
+    private static final String QUESTIONS_FIND_QUESTIONS = "questions/find-questions";
 
     private final QuestionService questionService;
 
@@ -20,7 +28,12 @@ public class QuestionController {
         this.questionService = questionService;
     }
 
-    @GetMapping({"", "/", "list", "list.html", "index", "index.html"})
+    @InitBinder
+    public void setAllowedFields(WebDataBinder dataBinder) {
+        dataBinder.setDisallowedFields("id");
+    }
+
+    @GetMapping({"list", "list.html", "index", "index.html"})
     public String list(Model model) {
         model.addAttribute("questions", questionService.findAll());
         return "questions/list";
@@ -35,8 +48,28 @@ public class QuestionController {
         return modelAndView;
     }
 
-    @RequestMapping("/find")
-    public String find() {
-        return "notimplemented";
+    @GetMapping("/find")
+    public String initFindQuestionForm(Model model) {
+        model.addAttribute(new Question());
+        return QUESTIONS_FIND_QUESTIONS;
+    }
+
+    @GetMapping("")
+    public String processFindQuestionForm(Question question, BindingResult result, Model model) {
+        if (question.getWordingEn() == null) {
+            question.setWordingEn("");
+        }
+        String searchPattern = "%" + question.getWordingEn() + "%";
+        Set<Question> questions = questionService.findAllByWordingEnLike(searchPattern);
+        if (questions.isEmpty()) {
+            result.rejectValue("wordingEn", "notFound", "not found");
+            return QUESTIONS_FIND_QUESTIONS;
+        } else if (questions.size() == 1) {
+            Question one = new ArrayList<>(questions).get(0);
+            return "redirect:/questions/" + one.getId();
+        } else {
+            model.addAttribute("questions", questions);
+            return "questions/list";
+        }
     }
 }